npu: release android ddk6.4.0.10 [2/3]
PD#SWPL-18595

Problem:
  DDK 6.4.0.10 Android porting

Solution:
  update nn source code base on driver6.4.0.10

Verify:
  local on sm1/w400

+++ applib/ovxinc/include/utils/vsi_nn_binary_tree.h
@@ -28,7 +28,9 @@
 extern "C"{
 #endif
 
-typedef long long vsi_nn_binary_tree_key_t;
+#include <stdint.h>
+
+typedef int64_t vsi_nn_binary_tree_key_t;
 
 #define vsi_nn_BinaryTreeInitRoot(n) do{n = NULL;} while (0);
